rename: assets 폴더를 생성하고, 기존의 images 파일을 해당 폴더 안으로 이동
