+++ src/pages/Mypage/components/AddIdolListItem.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import './AddIdolListItem.css';
-import checkIcon from '../images/ic_check.svg';
+import checkIcon from '../assets/images/ic_check.svg';
 
 function AddIdolListItem({ isSelected, idolData }) {
   const { name, group, profilePicture } = idolData;
